# swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "version": "1.0.0",
      "description": "Social Media API's to register and authenticate a user, create post and other operations assocated with post",
      "title": "Postaway"
    },
    "servers": [
      {
        "url": "http://localhost:3500"
      }
    ],
    "components": {
      "securitySchemes": {
        "JWT": {
          "type": "apiKey",
          "in": "header",
          "name": "Authorization"
        }
      }
    },
    "paths": {
...
      "/api/posts": {
        "post": {
          "tags": [
            "Posts"
          ],
          "summary": "Create a post",
          "security": [
            {
              "JWT": []
            }
          ],
          "requestBody": {
            "content": {
              "multipart/form-data": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "caption": {
                      "type": "string"
                    },
                    "imageUrl": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                }
              }
            },
            "description": "Enter Caption and file to create a post"
          },
          "responses": {
            "201": {
              "description": "Created"
            },
            "400": {
              "description": "Bad request"
            }
          }
        },
        "get": {
          "tags": [
            "Posts"
          ],
          "summary": "Get user posts",
          "description": "Get post based on user credentials",
          "security": [
            {
              "JWT": []
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      },
...
    }
  }